Add boot migrator
