+++ src/main/java/de/terrestris/shogun/migrator/util/MigrationException.java
@@ -1,0 +1,9 @@
+package de.terrestris.shogun.migrator.util;
+
+public class MigrationException extends RuntimeException {
+
+  public MigrationException(Throwable cause) {
+    super("Unable to migrate", cause);
+  }
+
+}
