+++ src/main/resources/META-INF/services/de.terrestris.shogun.migrator.spi.ShogunMigrator
@@ -1,0 +1,2 @@
+de.terrestris.shogun.migrator.shogun2.Shogun2Migrator
+de.terrestris.shogun.migrator.shogun2.BootMigrator
