+++ src/site/site.xml
@@ -1,0 +1,11 @@
+<project name="SHOGun migrator" combine.self="override">
+  <skin>
+    <groupId>org.apache.maven.skins</groupId>
+    <artifactId>maven-fluido-skin</artifactId>
+    <version>1.9</version>
+  </skin>
+  <body>
+    <menu ref="modules" inherit="top"/>
+    <menu ref="reports" inherit="top"/>
+  </body>
+</project>
