+++ src/test/java/de/terrestris/shogun/migrator/MigratorTest.java
@@ -2,17 +2,14 @@ package de.terrestris.shogun.migrator;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import de.terrestris.shogun.migrator.shogun2.Shogun2Migrator;
 import org.apache.commons.io.IOUtils;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import java.io.ByteArrayInputStream;
-import java.io.File;
 import java.io.IOException;
 import java.util.HashMap;
-
-import static org.apache.commons.io.FileUtils.copyInputStreamToFile;
 
 public class MigratorTest {
 
@@ -22,7 +19,7 @@ public class MigratorTest {
     @ValueSource(strings = {"/1.json", "/2.json", "/3.json"})
     public void testMigration(String file) throws IOException {
         JsonNode node = mapper.readTree(MigratorTest.class.getResource(file));
-        byte[] bs = Migrator.migrateApplication(node, new HashMap<>());
+        byte[] bs = Shogun2Migrator.migrateApplication(node, new HashMap<>());
         byte[] expected = IOUtils.toByteArray(MigratorTest.class.getResource("/migrated" + file));
         Assertions.assertArrayEquals(expected, bs);
     }
@@ -31,7 +28,7 @@ public class MigratorTest {
     @ValueSource(strings = {"/layer1.json", "/layer2.json", "/layer3.json", "/layer4.json", "/layer5.json", "/layer6.json", "/layer7.json", "/layer8.json"})
     public void testLayerMigration(String file) throws IOException {
         JsonNode node = mapper.readTree(MigratorTest.class.getResource(file));
-        byte[] bs = Migrator.migrateLayer(node);
+        byte[] bs = Shogun2Migrator.migrateLayer(node);
         byte[] expected = IOUtils.toByteArray(MigratorTest.class.getResource("/migratedlayer" + file));
         Assertions.assertArrayEquals(expected, bs);
     }
